Separate problem context from definition
Currently, the Problems API does not differentiate between the data defining a problem and the contextual information (ie the data associated with the occurrences). This pull request contains the first set of changes for the separation. It modifies only the internal API as follows
-  `org.gradle.api.problems.internal.Problem` is renamed to  `org.gradle.api.problems.internal.ProblemReport`
- ProblemReport now has only 2 fields for retrieving data: `getProblemDefinition()`, `getProblemContext()`.

The change allows us to:
- express separation of problem definition and context for all clients (TAPI, and other `ProblemProgressDetails` consumers)
- enhance `Problems` service and declare what constitutes a problem definition and what is part of the context.
- introduce deduplication strategies on exposed TAPI problem events.

As a follow-up, I'll attempt to add more capabilities to the (internal) ProblemSpec to define more contextual and non-contextual data, such as `label` and `contextualLabel`.